# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Coopuertos",
        "description": "API REST para el sistema Coopuertos - Gestión de conductores, vehículos y propietarios",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor API v1"
        }
    ],
    "paths": {
...
        "/api/v1/propietarios": {
            "get": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Listar propietarios",
                "description": "Retorna una lista paginada de propietarios con filtros opcionales",
                "operationId": "b5dd90e2f7c19f8cfd22463f6a2cd514",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de propietarios"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Crear propietario",
                "description": "Crea un nuevo propietario",
                "operationId": "101b99dc0b8bfb40c4820318f22c94d4",
                "responses": {
                    "201": {
                        "description": "Propietario creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/propietarios/{id}": {
            "get": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Mostrar propietario",
                "description": "Retorna la información de un propietario específico",
                "operationId": "d32a49f20a2f1bb49dfb6a6c1334bf58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del propietario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del propietario"
                    },
                    "404": {
                        "description": "Propietario no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Actualizar propietario",
                "description": "Actualiza la información de un propietario",
                "operationId": "cd7bc35f4eb262160f25359e1a301cbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del propietario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propietario actualizado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Propietario no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Eliminar propietario",
                "description": "Elimina un propietario del sistema",
                "operationId": "f5dfca5d1e6b03dc306b30a5c3dd73ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del propietario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propietario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Propietario no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/propietarios/search": {
            "get": {
                "tags": [
                    "Propietarios"
                ],
                "summary": "Buscar propietarios",
                "description": "Busca propietarios por nombre completo o número de identificación",
                "operationId": "e0082ee23785aa3a19ce97f1d7e020ce",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación de Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}